pip_requirements.txt
# Usage: pip install -r cuda11_pip_requirements.txt
# for python 3.7 or 3.8, cuda > 11
-f https://download.pytorch.org/whl/lts/1.8/torch_lts.html
joblib==1.1.0
pandas==1.2.5
numpy==1.21.4
Pillow==9.0.0
python-dateutil==2.8.2
pytz==2021.1
scikit-learn==0.24.2
scipy==1.7.0
matplotlib
torch==1.8.2+cu111
torchvision==0.9.2+cu111
einops==0.3.0
utm
git+https://github.com/facebookresearch/fvcore
timm
psutil
simplejson
transformers
faiss-cpu==1.7.1.post3